requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
typing-extensions==4.1.1
